+++ src/eca/server.clj
@@ -45,9 +45,6 @@
 (defmethod lsp.server/receive-request "chat/queryContext" [_ components params]
   (handlers/chat-query-context components params))
 
-(defmethod lsp.server/receive-request "mcp/listServers" [_ components params]
-  (handlers/mcp-list-servers components params))
-
 (defn ^:private monitor-server-logs [log-ch]
   ;; NOTE: if this were moved to `initialize`, after timbre has been configured,
   ;; the server's startup logs and traces would appear in the regular log file
@@ -77,6 +74,9 @@
   (chat-content-received [_this content]
     (lsp.server/discarding-stdout
      (lsp.server/send-notification server "chat/contentReceived" content)))
+  (mcp-server-updated [_this params]
+    (lsp.server/discarding-stdout
+     (lsp.server/send-notification server "mcp/serverUpdated" params)))
   (showMessage [_this msg]
     (lsp.server/discarding-stdout
      (lsp.server/send-notification server "$/showMessage" msg))))